+++ lib/require_profiler/printer/call_stack.rb
@@ -5,7 +5,7 @@ module RequireProfiler
     # CallStack formatter prints collapsed stacks (Brendan Gregg's format)
     class CallStack < Base
       def flush(node, parts: [])
-        return unless (node.time * 1000) >= threshold
+        return unless flush?(node)
 
         path = node.path.sub(prefix_stripper, "")
         self_parts = path.split("/")