+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+- Add focus support (to include only matching files). ([@palkan][])
+
 - Add threshold support (`REQUIRE_PROFILE_THRESHOLD`). ([@palkan][])
 
 - Add `-require-prof` support. ([@palkan][])